requirements.txt
accelerate==0.15.0
bitsandbytes==0.35.4
decord==0.6.0
diffusers[torch]==0.11.1
einops==0.6.0
ftfy==6.1.1
gradio==3.16.2
huggingface-hub==0.12.0
imageio==2.25.0
imageio-ffmpeg==0.4.8
omegaconf==2.3.0
Pillow==9.4.0
python-slugify==7.0.0
tensorboard==2.11.2
torch==1.13.1
torchvision==0.14.1
transformers==4.26.0
triton==2.0.0.dev20221202
xformers==0.0.16
